+++ matlab/TargetRobot.m
@@ -2,7 +2,7 @@ classdef TargetRobot < Roomba
     
     properties
         state; % [x, y, theta, timer]
-        rotating; % delta T value reflecting how long it takes for the robot to complete rotatation. Not rotation if set to 0
+        rotating; % (double) delta T value reflecting how long it takes for the robot to complete rotatation. Not rotation if set to 0
         collidedThisTimeStep = false;
         roombaID;
     end
